Extract src/taskrun.ts — run on stop, and a run's aftermath
Step 2 of the bootstrap trim, and PLAN's prediction for it held: moving this
block *removed* a host object rather than adding one. inspector.ts had a
three-member host solely because rerunTask / revealSource /
sendOutputToSession lived in main.ts; they are now below it, so
setInspectorHost is gone and it imports them directly — the same win the
${input:…} move got.

Moved (four guarded ranges): STOP_RUN_FLOOR + stopRunAt + stopInFlight and
maybeRunOnStop, then rerunTask, sendOutputToSession and revealSource.

launchTask deliberately stayed behind. It builds a pane — xterm, addons, the
PTY spawn, a `Sess` — which is the pane layer, item 3 in PLAN's order, so it
is a setTaskRunLaunchTask seam until then rather than being dragged into a
module about *when* to run something.

Three seam-rule-2 setters, all pane operations: setTaskRunSetActive,
setTaskRunCloseSession, setTaskRunLaunchTask.

Verified mechanical: line-sorted against `git show HEAD:src/main.ts` the
relocated block is identical modulo four `export` keywords. Block order
differs on purpose — run-on-stop leads, since it is what the module is for.

main.ts 1,247 -> 1,169. Gates: tsc --noEmit clean (exit 0), 368 vitest pass.
No Rust touched.
