+++ src/main.ts
@@ -22,7 +22,10 @@ import { closePalette, openPalette, setPaletteHost } from "./palui";
 import {
   closeColorPop, closeCtxMenu, ctxMenuOpen, openColorPopover, setProjMenuHost,
 } from "./projmenu";
-import { renderInspector, setInspectorHost } from "./inspector";
+import { renderInspector } from "./inspector";
+import {
+  maybeRunOnStop, setTaskRunCloseSession, setTaskRunLaunchTask, setTaskRunSetActive,
+} from "./taskrun";
 import {
   closeExternalView, flushRoster, forgetDormant, jumpExternal, loadDormants,
   openDormant, openExternal, queueRosterSave, refreshDirtyStates, refreshExternals,
@@ -52,8 +55,8 @@ import { setRlLogger } from "./rl";
 import { closeCafPop, reconcileCaf, setCafHost } from "./caffeinate";
 import { closeDiff, diffOpen, openDiff, setDiffCloseFootMenus } from "./diffview";
 import {
-  closeInputPrompt, closeRunPicker, closeTaskManager, mgrEdit, openInputPrompt,
-  openRunPicker, renderMgr, setMgrEdit, setTaskUiHost,
+  closeInputPrompt, closeRunPicker, closeTaskManager, mgrEdit, openRunPicker,
+  renderMgr, setMgrEdit, setTaskUiHost,
 } from "./taskui";
 import {
   closeSettings, openSettings, renderSettings, setSettingsHost, setTab, settingsOpen,
@@ -74,9 +77,8 @@ import {
   nextAfterClose, orderedSessions,
 } from "./grouping";
 import {
-  discoverTasks, execCmd, exitWaiters, lastRunnableById, launchWithDeps,
-  setTaskLauncher, setTaskLogger, setTaskRepaint, setTaskToast, stopRuleBlocked,
-  stopRules, taskPrefs, type TaskLaunchOpts,
+  execCmd, exitWaiters, setTaskLauncher, setTaskLogger, setTaskRepaint, setTaskToast,
+  taskPrefs, type TaskLaunchOpts,
 } from "./tasks";
 
 // One-time recovery of localStorage stranded when the app was renamed
@@ -199,9 +201,10 @@ setProjMenuHost({
   renderAll, requestLaunch, launchShell, openProjectFolder, addProjectPath,
   removeFavorite,
 });
-// The task card's actions: re-run, reveal the file it came from, hand the failure to
-// an agent. All three own panes or the backend, so none of them is the inspector's.
-setInspectorHost({ rerunTask, revealSource, sendOutputToSession });
+// Run-on-stop and the task inspector's actions reach back for three pane operations.
+setTaskRunSetActive(setActive);
+setTaskRunCloseSession(closeSession);
+setTaskRunLaunchTask(launchTask);
 // The read-only mirrors reconcile the stage when what they point at goes away, and a
 // dormant row resumes into a real pane — neither of which they own.
 setMirrorSetActive(setActive);
@@ -526,77 +529,18 @@ function renderHeader(s: Sess | null) {
 // moved to ./tasks, beside the runner override and remembered inputs it already had.
 
 // ---------- runnables (tasks & scripts) ----------
-// Discovery lives in Rust (src-tauri/src/tasks.rs) and only ever *parses* files —
-// it never executes the project to find out what it can do. This half is choosing
-// and observing.
+// Discovery lives in Rust (src-tauri/src/tasks.rs) and only ever *parses* files — it
+// never executes the project to find out what it can do. Of the frontend half, the
+// decisions are in ./tasks, the two surfaces in ./taskui, and run-on-stop plus the
+// task inspector's actions in ./taskrun. What stays here is launchTask: it opens a
+// pane, which is this file's job.
 
 
 // The package-runner override and the remembered ${input:…} values now live in
 // ./tasks, beside the substitution that consumes them.
 
-// ↗ Reveal source — where a task came from, selected in the OS file manager. `root`
-// is the directory discovery ran in, so the repo-relative `sourceFile` resolves; a
-// blocked/synthetic row has no real file and shows nothing to reveal.
-function revealSource(root: string, sourceFile: string) {
-  invoke("reveal_path", { dir: root, rel: sourceFile }).catch((e) => toast("reveal failed: " + e));
-}
 
 
-// A Stop fires at the end of *every* turn — that's the point — but two can land in
-// quick succession, and a slow suite must never race a second copy of itself in
-// the same worktree. The floor is deliberately short: it exists to swallow a
-// double-fire, not to ration runs.
-const stopRunAt = new Map<string, number>();
-const STOP_RUN_FLOOR = 5000;
-// A run-on-stop launch is only visible as a pane *after* its dependency chain has
-// run — so a rule with `dependsOn` has no `run.id === rule.id` pane during the whole
-// dep phase, and the 5s floor alone can't stop a turn that lands mid-build from
-// racing a second chain. This marks "a chain for this project is starting", claimed
-// synchronously before the first await and cleared once the launch settles; by then
-// the rule pane exists and the in-flight scan below takes over.
-const stopInFlight = new Set<string>();
-
-async function maybeRunOnStop(s: Sess) {
-  const rule = stopRules[s.colorKey];
-  if (!rule || !isAgent(s)) return;
-  // Claimed before the first await: discovery is async, so two Stops in the same
-  // tick would otherwise both get past this.
-  if (Date.now() - (stopRunAt.get(s.colorKey) ?? 0) < STOP_RUN_FLOOR) return;
-  // A chain still starting (deps running, rule pane not created yet) wins — the pane
-  // scan below can't see it, so this covers the window the floor can't.
-  if (stopInFlight.has(s.colorKey)) {
-    dlog("info", `run-on-stop ${rule.id} skipped — a chain is already starting`);
-    return;
-  }
-  // A run of this rule still in flight wins. Restarting the suite from the top
-  // mid-flight tells you nothing and doubles the load on the machine.
-  if ([...sessions.values()].some((x) => x.kind === "task" && x.colorKey === s.colorKey && x.run?.id === rule.id && x.run.exitCode == null)) {
-    dlog("info", `run-on-stop ${rule.id} skipped — still running`);
-    return;
-  }
-  stopRunAt.set(s.colorKey, Date.now());
-  stopInFlight.add(s.colorKey);
-  try {
-    // Discover in the *session's* workdir, so with several worktrees of one repo
-    // open the run verifies the checkout the agent actually just edited. Hidden
-    // tasks count — hiding is about the picker, not about what may run.
-    const spec = (await discoverTasks(s.workdir, s.colorKey, true)).find((r) => r.id === rule.id);
-    if (!spec) {
-      dlog("warn", `run-on-stop ${rule.id} gone from ${s.project}`);
-      toast(`Run after a turn: “${rule.label}” isn’t in ${s.project} any more`);
-      return;
-    }
-    const why = stopRuleBlocked(spec);
-    if (why) { dlog("warn", `run-on-stop ${rule.id} skipped: ${why}`); return; }
-    dlog("info", `run-on-stop ${rule.id} · ${s.project} · ${s.id.slice(0, 8)} finished a turn`);
-    await launchWithDeps(spec, s.project, {
-      colorKey: s.colorKey, worktree: s.worktree, branch: s.branch,
-      discoveredIn: spec.cwd, forSession: s.id, focus: false,
-    });
-  } finally {
-    stopInFlight.delete(s.colorKey);
-  }
-}
 
 // "a, b and c" — the quit guard lists up to three kinds of running thing.
 function listPhrase(parts: string[]): string {
@@ -604,20 +548,6 @@ function listPhrase(parts: string[]): string {
   return `${parts.slice(0, -1).join(", ")} and ${parts[parts.length - 1]}`;
 }
 
-// Type the failure into a Claude session's stdin — deliberately *without* a
-// trailing newline, so you read what's about to be sent and press Enter yourself.
-// Same contract as handToTerminal: Episko prefills, the human commits.
-function sendOutputToSession(task: Sess, targetId: string) {
-  const t = sessions.get(targetId);
-  if (!t?.term) { toast("That session is gone"); return; }
-  const r = task.run!;
-  const tail = r.tail.join("\n").trim();
-  const msg = `\`${r.cmd}\` failed with exit ${r.exitCode}:\n\n${tail}\n\nPlease fix it.`;
-  setActive(targetId);
-  invoke("write_pty", { sessionId: targetId, data: msg.replace(/\n/g, "\r") })
-    .then(() => toast("Pasted into the session — press Enter to send"))
-    .catch((e) => toast("send failed: " + e));
-}
 
 // Start one run of a Runnable in its own pane. Mirrors launchShell — same PTY,
 // same xterm setup — because a task genuinely is just another pane.
@@ -685,18 +615,6 @@ async function launchTask(r: Runnable, project: string, opts: TaskLaunchOpts = {
 // The dependsOn chain — resolveDeps / launchWithDeps / waitForExit — now lives in
 // ./tasks, which reaches launchTask below through setTaskLauncher.
 
-// Re-running reuses nothing — it opens a fresh pane and closes the old one, so the
-// sidebar doesn't accumulate a row per attempt while the scrollback stays honest
-// about which attempt you're reading.
-async function rerunTask(s: Sess) {
-  const r = s.run; if (!r) return;
-  const spec = lastRunnableById.get(r.id);
-  if (!spec) { toast("Task definition is gone — rescan"); return; }
-  if (spec.inputs.length) { openInputPrompt(spec, s.project, { colorKey: s.colorKey, worktree: s.worktree, branch: s.branch, discoveredIn: spec.cwd }); return; }
-  const project = s.project, colorKey = s.colorKey, worktree = s.worktree, branch = s.branch;
-  closeSession(s.id);
-  await launchTask(spec, project, { colorKey, worktree, branch });
-}
 
 // Discovery (discoverTasks / rescanTasks) moved to ./tasks — every input they read
 // now lives there.
